backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic[email]>=2.8.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
azure-cosmos>=4.5.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
python-dotenv>=1.0.0
psutil>=5.9.0
redis>=5.0.0
email-validator>=2.0.0